==> 1-load_data.Rmd <==
---
title: "1-load_data.Rmd"
author: "Antton Alberdi"
date: "2023-12-12"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load libraries

```{r libraries}
library(tidyverse)
library(ape)
library(distillR)
```

## Retrieve and load data

### Count table
This is the document containing the number of sequencing reads from each sample have been mapped to each MAG. Note that this is the raw data that needs to be further processed before running any statistics on them.

```{r load_count, message=FALSE}
read_counts <- read_tsv("https://sid.erda.dk/share_redirect/gR0yYR6oKL/results/stats/coverm_genome_2023-11-03_swine_mannan_mg.count.tsv") %>%
  rename_all(~ str_remove_all(., " Read Count")) %>% #simplify column names
  rename(genome = 1)

genomes <- read_counts$genome # create list of genome names
```

### Base hit table
This is the document containing the number of nucleotide bases have been covered by at least one read in each sample and MAG. This information is used to calculate MAG coverage values.

```{r load_hits, message=FALSE}
basehits <- read_tsv("https://sid.erda.dk/share_redirect/gR0yYR6oKL/results/stats/coverm_genome_2023-11-03_swine_mannan_mg.covered_bases.tsv") %>%
  rename_all(~ str_remove_all(., " Covered Bases")) %>% #simplify column names
  rename(genome = 1) %>%
  arrange(match(genome, genomes))
```

### Sample metadata

```{r load_sample_metadata, message=FALSE}
sample_metadata <- read_csv("data/sample_metadata.csv") %>%
  mutate(type=ifelse(grepl("FD", library), "digesta", ifelse(grepl("MD", library), "faeces", NA))) #add sample type column based on library name
```

### Genome metadata

#### Taxonomy
This is the raw taxonomy table generated by GTDBtk, which is simplified for downstream analyses.
```{r load_taxonomy, message=FALSE}
genome_taxonomy <- read_tsv("https://sid.erda.dk/share_redirect/BLJUHsGfWu/results/dereplicate/gtdbtk/gtdbtk.summary.tsv") %>%
  rename(genome = user_genome) %>%
  mutate(genome = str_replace_all(genome,"\\.fa", "")) %>%
  separate(classification, c("domain","phylum","class","order","family","genus","species"),  sep =";") %>%
  select(genome,domain,phylum,class,order,family,genus,species) %>%
  arrange(match(genome, genomes))
```

#### Genome lengths
Number of nucleotides of each genome. This information is needed to calculate coverage values and to normalise read counts by genome length.
```{r load_lengths, message=FALSE}
genome_lengths <- read_tsv("https://sid.erda.dk/share_redirect/gR0yYR6oKL/results/stats/coverm_genome_2023-11-03_swine_mannan_mg.length.tsv") %>%
  rename(genome = 1, length=2) %>% 
  select(genome,length) %>% #remove redundancy
  arrange(match(genome, genomes))
```

#### Genome quality
Quality properties of the genomes. 
```{r load_lengths, message=FALSE}
genome_quality <- read_csv("https://sid.erda.dk/share_redirect/BLJUHsGfWu/results/dereplicate/drep/data_tables/genomeInfo.csv") %>%
  rename(genome = 1) %>%
  mutate(genome = str_replace_all(genome,"\\.fa", "")) %>% #remove .fa from genome names
  filter(genome %in% genomes) %>%
  arrange(match(genome, genomes))
```

#### Merged metadata object
Merge taxonomy, length and quality information
```{r create_genomemetadata, message=FALSE}
genome_metadata <- genome_taxonomy %>%
  #left_join(genomelengths,by=join_by(genome==genome)) %>% #join lengths - not used because Genome quality already contains this information
  left_join(genomequality,by=join_by(genome==genome)) #join quality
```

### Genome tree
This is the raw tree generated by GTDBtk, which needs to be pruned to obtain the phylogenetic tree of the MAGs.
```{r load_tree}
archaea_tree <- read.tree(url("https://sid.erda.dk/share_redirect/BLJUHsGfWu/results/dereplicate/gtdbtk/classify/gtdbtk.ar53.classify.tree")) #note that when no archaea are detected, this tree is not generated
bacteria_tree <- read.tree(url("https://sid.erda.dk/share_redirect/BLJUHsGfWu/results/dereplicate/gtdbtk/classify/gtdbtk.backbone.bac120.classify.tree"))
genome_tree <- bind.tree(archaea_tree, bacteria_tree)
genome_tree$tip.label <- str_replace_all(genome_tree$tip.label,"'", "") #remove single quotes in MAG names
genome_tree <- keep.tip(genome_tree, tip=genomes) # keep only MAG tips
```

### MAG functional annotations

This is the raw annotation table generated by DRAM, which is used to generate GIFT data using distillR.
```{r load_annotations, message=FALSE}
genome_annotations <- read_tsv("https://sid.erda.dk/share_redirect/BLJUHsGfWu/results/dereplicate/dram/annotations.tsv") %>%
  rename(gene=1,genome=2)
```

## Transform and normalise data

### Generate coverage table
By dividing the number of base hits by the length of each MAG, coverage values can be calculated.

```{r calc_coverage}
genome_coverage <- basehits %>%
  mutate(across(where(is.numeric), ~ ./genome_metadata$length))
```

### Coverage filtering

```{r filter_coverage}
min_coverage=0.3
read_counts_filt <- genome_coverage %>%
  mutate(across(where(is.numeric), ~ ifelse(. > min_coverage, 1, 0))) %>%
  mutate(across(-1, ~ . * read_counts[[cur_column()]])) 
```

### Generate genome count table

```{r calc_genometable}
readlength=150 #change if sequencing read length is different
genome_counts <- read_counts_filt %>%
  mutate(across(where(is.numeric), ~ . / (genome_metadata$length / readlength) ))
```

### Distil functional annotations

```{r distil_annotations}
genome_gifts <- distill(genome_annotations,GIFT_db,genomecol=2,annotcol=c(9,10,19))
```

## Wrap working objects
In the last step, the objects that are needed for downstream analyses are stored in an R object.

```{r wrap_objects}
save(read_counts, genome_counts, genome_tree, genome_metadata, genome_gifts, sample_metadata, file = "data/data.Rdata")
```

- **read_counts**: Number of reads mapped to each genome in each sample. Note this is the unfiltered and unnormalised raw community composition table.
- **genome_counts**: Number of genomes quantified in each sample, calculated through filtering and normalising ***read_counts***. This is the community composition table to be used in downstream analyses unless otherwise stated.
- **genome_tree**: Phylogenetic tree of the genomes, to be employed in downstream phylogenetic analyses.
- **genome_metadata**: Taxonomic and quality information of the genomes.
- **genome_gifts**: Genome-inferred functional traits of the genomes, to be employed in downstream functional analyses.
- **sample_metadata**: Treatment/population and other relevant metadata of the samples.
